test: move theme tests from src into tests/
The inline #[cfg(test)] module in src/theme.rs was the only place tests
lived alongside code. Move the public-API tests to tests/theme.rs to match
the rest of the suite.

Drop the from_raw/RawTheme tests: those items are pub(crate) and their
behavior is already covered end-to-end through Config::load_from_file in
tests/config.rs.
