+++ src/theme.rs
@@ -251,131 +251,3 @@ mod color_serde {
         Color::from_str(&s).map_err(|_| D::Error::custom(format!("invalid color: '{s}'")))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn default_theme_round_trips_through_toml() {
-        let theme = Theme::default();
-        let toml_str = toml::to_string(&theme).expect("serialize theme");
-        let parsed: Theme = toml::from_str(&toml_str).expect("deserialize theme");
-        assert_eq!(theme, parsed);
-    }
-
-    #[test]
-    fn accepts_hex_colors() {
-        let toml_str = r##"due_date = "#ff8000""##;
-        let theme: Theme = toml::from_str(toml_str).expect("deserialize theme");
-        assert_eq!(theme.due_date, Color::Rgb(0xff, 0x80, 0x00));
-    }
-
-    #[test]
-    fn rejects_invalid_colors() {
-        let toml_str = r#"accent = "not-a-color""#;
-        let result: Result<Theme, _> = toml::from_str(toml_str);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn from_raw_defaults_missing_fields_silently() {
-        let source = "";
-        let raw: RawTheme = toml::from_str(source).unwrap();
-        let (theme, warnings) = Theme::from_raw(raw, source);
-        assert_eq!(theme, Theme::default());
-        assert!(warnings.is_empty());
-    }
-
-    #[test]
-    fn from_raw_keeps_valid_overrides_and_defaults_the_rest() {
-        let source = "accent = \"Magenta\"\n";
-        let raw: RawTheme = toml::from_str(source).unwrap();
-        let (theme, warnings) = Theme::from_raw(raw, source);
-        assert_eq!(theme.accent, Color::Magenta);
-        assert_eq!(theme.danger, Theme::default().danger);
-        assert!(warnings.is_empty());
-    }
-
-    #[test]
-    fn from_raw_falls_back_and_warns_on_invalid_color_string() {
-        let source = "accent = \"Blue\"\ndanger = \"notacolor\"\n";
-        let raw: RawTheme = toml::from_str(source).unwrap();
-        let (theme, warnings) = Theme::from_raw(raw, source);
-
-        assert_eq!(theme.accent, Color::Blue);
-        assert_eq!(
-            theme.danger,
-            Theme::default().danger,
-            "invalid value should fall back to default"
-        );
-
-        assert_eq!(warnings.len(), 1);
-        assert_eq!(warnings[0].field, "danger");
-        assert_eq!(warnings[0].raw_value, "notacolor");
-        assert_eq!(warnings[0].line, 2);
-    }
-
-    #[test]
-    fn from_raw_falls_back_and_warns_on_wrong_type() {
-        let source = "accent = 5\n";
-        let raw: RawTheme = toml::from_str(source).unwrap();
-        let (theme, warnings) = Theme::from_raw(raw, source);
-
-        assert_eq!(theme.accent, Theme::default().accent);
-        assert_eq!(warnings.len(), 1);
-        assert_eq!(warnings[0].field, "accent");
-        assert_eq!(warnings[0].line, 1);
-    }
-
-    #[test]
-    fn from_raw_reports_multiple_warnings_with_correct_lines() {
-        let source = "accent = \"Blue\"\ndanger = \"bogus1\"\nborder = \"bogus2\"\n";
-        let raw: RawTheme = toml::from_str(source).unwrap();
-        let (_, warnings) = Theme::from_raw(raw, source);
-
-        assert_eq!(warnings.len(), 2);
-        assert_eq!(warnings[0].field, "danger");
-        assert_eq!(warnings[0].line, 2);
-        assert_eq!(warnings[1].field, "border");
-        assert_eq!(warnings[1].line, 3);
-    }
-
-    #[test]
-    fn format_warnings_returns_none_when_empty() {
-        assert_eq!(format_warnings(&[]), None);
-    }
-
-    #[test]
-    fn format_warnings_summarizes_all_entries() {
-        let warnings = vec![
-            ThemeWarning {
-                field: "danger",
-                raw_value: "notacolor".to_string(),
-                line: 2,
-            },
-            ThemeWarning {
-                field: "border",
-                raw_value: "bogus".to_string(),
-                line: 3,
-            },
-        ];
-        let message = format_warnings(&warnings).unwrap();
-        assert!(message.contains("2 theme colors"));
-        assert!(message.contains("theme.danger (line 2)"));
-        assert!(message.contains("theme.border (line 3)"));
-    }
-
-    #[test]
-    fn theme_warning_display_is_readable() {
-        let warning = ThemeWarning {
-            field: "danger",
-            raw_value: "notacolor".to_string(),
-            line: 2,
-        };
-        assert_eq!(
-            warning.to_string(),
-            "theme.danger (line 2): 'notacolor' is not a valid color, using default"
-        );
-    }
-}
